refactor(ScrollableNav): Move next to Masthead
