+++ packages/react-component-library/src/components/TopLevelNavigation/Masthead/MastheadNav.tsx
@@ -4,7 +4,7 @@ import { Nav, NavItem } from '../../../types/Nav'
 import {
   ScrollableNav,
   ScrollableNavItem,
-} from "../../ScrollableNav"
+} from "../ScrollableNav"
 
 export const MastheadNav: React.FC<Nav<NavItem>> = props => {
   return (
